+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iSunFA",
-  "version": "0.8.5+115",
+  "version": "0.8.5+116",
   "private": false,
   "scripts": {
     "dev": "next dev",
