+++ src/components/voucher/new_voucher_form.tsx
@@ -5,7 +5,7 @@ import { BiSave } from 'react-icons/bi';
 import { FiSearch } from 'react-icons/fi';
 import { useTranslation } from 'next-i18next';
 import useOuterClick from '@/lib/hooks/use_outer_click';
-import { useHotkeys } from 'react-hotkeys-hook';
+// import { useHotkeys } from 'react-hotkeys-hook';
 import { Button } from '@/components/button/button';
 import DatePicker, { DatePickerType } from '@/components/date_picker/date_picker';
 // import Toggle from '@/components/toggle/toggle';
@@ -53,7 +53,7 @@ import { IAIResultVoucher } from '@/interfaces/voucher';
 //   WEEK = 'week',
 // }
 
-type FocusableElement = HTMLInputElement | HTMLButtonElement | HTMLDivElement;
+// type FocusableElement = HTMLInputElement | HTMLButtonElement | HTMLDivElement;
 
 const dummyAIResult: IAIResultVoucher = {
   voucherDate: 0,
@@ -342,111 +342,111 @@ const NewVoucherForm: React.FC<NewVoucherFormProps> = ({ selectedData }) => {
   // Info: (20241107 - Julian) ============ 熱鍵設置 ============
   const formRef = useRef<HTMLFormElement>(null);
 
-  const handleTabPress = useCallback(
-    (event: KeyboardEvent) => {
-      event.preventDefault(); // Info: (20241107 - Julian) 阻止預設事件
+  // const handleTabPress = useCallback(
+  //   (event: KeyboardEvent) => {
+  //     event.preventDefault(); // Info: (20241107 - Julian) 阻止預設事件
 
-      // Info: (20241107 - Julian) 獲取 form 元素中的所有 input, button 元素
-      const elementsInForm =
-        formRef.current?.querySelectorAll<FocusableElement>('input, button, div') ?? [];
+  //     // Info: (20241107 - Julian) 獲取 form 元素中的所有 input, button 元素
+  //     const elementsInForm =
+  //       formRef.current?.querySelectorAll<FocusableElement>('input, button, div') ?? [];
 
-      // Info: (20241107 - Julian) 過濾出可聚焦的元素
-      const focusableElements: FocusableElement[] = Array.from(elementsInForm).filter(
-        // Info: (20241107 - Julian) 過濾掉 disabled 或 tabIndex < 0 的元素
-        (el) => el.tabIndex >= 0 && (el as HTMLInputElement | HTMLButtonElement).disabled !== true
-      );
+  //     // Info: (20241107 - Julian) 過濾出可聚焦的元素
+  //     const focusableElements: FocusableElement[] = Array.from(elementsInForm).filter(
+  //       // Info: (20241107 - Julian) 過濾掉 disabled 或 tabIndex < 0 的元素
+  //       (el) => el.tabIndex >= 0 && (el as HTMLInputElement | HTMLButtonElement).disabled !== true
+  //     );
 
-      // Info: (20241107 - Julian) 獲取各個欄位的 index
-      const dateIndex = focusableElements.findIndex((el) => el.id === 'voucher-date');
-      const voucherTypeIndex = focusableElements.findIndex((el) => el.id === 'voucher-type');
-      const noteIndex = focusableElements.findIndex((el) => el.id === 'voucher-note');
-      const counterpartyIndex = focusableElements.findIndex(
-        (el) => el.id === 'voucher-counterparty'
-      ); // Info: (20241108 - Julian) Div
-      const assetIndex = focusableElements.findIndex((el) => el.id === 'voucher-asset');
-      const accountTitleIndex = focusableElements.findIndex((el) =>
-        el.id.includes('account-title')
-      ); // Info: (20241108 - Julian) Div
+  //     // Info: (20241107 - Julian) 獲取各個欄位的 index
+  //     const dateIndex = focusableElements.findIndex((el) => el.id === 'voucher-date');
+  //     const voucherTypeIndex = focusableElements.findIndex((el) => el.id === 'voucher-type');
+  //     const noteIndex = focusableElements.findIndex((el) => el.id === 'voucher-note');
+  //     const counterpartyIndex = focusableElements.findIndex(
+  //       (el) => el.id === 'voucher-counterparty'
+  //     ); // Info: (20241108 - Julian) Div
+  //     const assetIndex = focusableElements.findIndex((el) => el.id === 'voucher-asset');
+  //     const accountTitleIndex = focusableElements.findIndex((el) =>
+  //       el.id.includes('account-title')
+  //     ); // Info: (20241108 - Julian) Div
 
-      const formIndexOrder = [
-        dateIndex,
-        voucherTypeIndex,
-        noteIndex,
-        counterpartyIndex,
-        assetIndex,
-        accountTitleIndex,
-      ];
+  //     const formIndexOrder = [
+  //       dateIndex,
+  //       voucherTypeIndex,
+  //       noteIndex,
+  //       counterpartyIndex,
+  //       assetIndex,
+  //       accountTitleIndex,
+  //     ];
 
-      // Info: (20241107 - Julian) 獲取當前聚焦元素的 index
-      const currentIndex = focusableElements.findIndex((el) => el === document.activeElement);
+  //     // Info: (20241107 - Julian) 獲取當前聚焦元素的 index
+  //     const currentIndex = focusableElements.findIndex((el) => el === document.activeElement);
 
-      const ToNext = () => {
-        // Info: (20241107 - Julian) 獲取下一個聚焦元素的 index
-        const nextIndex = currentIndex + 1 >= focusableElements.length ? 0 : currentIndex + 1;
-        // Info: (20241107 - Julian) 移動到下一個可聚焦元素
-        focusableElements[nextIndex]?.focus();
-      };
+  //     const ToNext = () => {
+  //       // Info: (20241107 - Julian) 獲取下一個聚焦元素的 index
+  //       const nextIndex = currentIndex + 1 >= focusableElements.length ? 0 : currentIndex + 1;
+  //       // Info: (20241107 - Julian) 移動到下一個可聚焦元素
+  //       focusableElements[nextIndex]?.focus();
+  //     };
 
-      // ToDo: (20241107 - Julian) ============ 施工中🔧 ============
-      if (currentIndex === -1 || currentIndex === focusableElements.length - 1) {
-        focusableElements[0]?.focus();
-      } else if (currentIndex >= formIndexOrder[0] && currentIndex < formIndexOrder[1]) {
-        // Info: (20241107 - Julian) 如果當前聚焦元素是日期欄位，且日期已選，則移動到類型欄位
-        if (date.startTimeStamp !== 0 && date.endTimeStamp !== 0) {
-          focusableElements[voucherTypeIndex]?.focus();
-        } else {
-          ToNext();
-        }
-      } else {
-        ToNext();
-      }
+  //     // ToDo: (20241107 - Julian) ============ 施工中🔧 ============
+  //     if (currentIndex === -1 || currentIndex === focusableElements.length - 1) {
+  //       focusableElements[0]?.focus();
+  //     } else if (currentIndex >= formIndexOrder[0] && currentIndex < formIndexOrder[1]) {
+  //       // Info: (20241107 - Julian) 如果當前聚焦元素是日期欄位，且日期已選，則移動到類型欄位
+  //       if (date.startTimeStamp !== 0 && date.endTimeStamp !== 0) {
+  //         focusableElements[voucherTypeIndex]?.focus();
+  //       } else {
+  //         ToNext();
+  //       }
+  //     } else {
+  //       ToNext();
+  //     }
 
-      // switch (currentIndex) {
-      //   case dateIndex:
-      //     if (date.startTimeStamp !== 0 && date.endTimeStamp !== 0) {
-      //       focusableElements[voucherTypeIndex]?.focus();
-      //     } else ToNext();
-      //     break;
-      //   case voucherTypeIndex:
-      //     console.log('voucherTypeIndex');
-      //     ToNext();
-      //     break;
-      //   case noteIndex:
-      //     console.log('noteIndex');
-      //     ToNext();
-      //     break;
-      //   case counterpartyIndex:
-      //     console.log('counterpartyIndex');
-      //     // Info: (20241107 - Julian) 如果需填入交易對象，但交易對象未選擇，則聚焦到交易對象欄位
-      //     if (isCounterpartyRequired && !counterparty) {
-      //       focusableElements[counterpartyIndex]?.click();
-      //     } else {
-      //       ToNext();
-      //     }
-      //     break;
-      //   case assetIndex:
-      //     console.log('assetIndex');
-      //     // Info: (20241107 - Julian) 如果需填入資產，但資產為空，則聚焦到資產欄位
-      //     if (isAssetRequired && temporaryAssetList.length === 0) {
-      //       focusableElements[assetIndex]?.focus();
-      //     } else {
-      //       ToNext();
-      //     }
-      //     break;
-      //   case accountTitleIndex:
-      //     console.log('accountTitleIndex');
-      //     ToNext();
-      //     break;
-      //   default:
-      //     console.log('default');
-      //     focusableElements[0]?.focus();
-      //     break;
-      // }
-    },
-    [formRef, date, counterparty, isCounterpartyRequired, temporaryAssetListByUser]
-  );
+  //     // switch (currentIndex) {
+  //     //   case dateIndex:
+  //     //     if (date.startTimeStamp !== 0 && date.endTimeStamp !== 0) {
+  //     //       focusableElements[voucherTypeIndex]?.focus();
+  //     //     } else ToNext();
+  //     //     break;
+  //     //   case voucherTypeIndex:
+  //     //     console.log('voucherTypeIndex');
+  //     //     ToNext();
+  //     //     break;
+  //     //   case noteIndex:
+  //     //     console.log('noteIndex');
+  //     //     ToNext();
+  //     //     break;
+  //     //   case counterpartyIndex:
+  //     //     console.log('counterpartyIndex');
+  //     //     // Info: (20241107 - Julian) 如果需填入交易對象，但交易對象未選擇，則聚焦到交易對象欄位
+  //     //     if (isCounterpartyRequired && !counterparty) {
+  //     //       focusableElements[counterpartyIndex]?.click();
+  //     //     } else {
+  //     //       ToNext();
+  //     //     }
+  //     //     break;
+  //     //   case assetIndex:
+  //     //     console.log('assetIndex');
+  //     //     // Info: (20241107 - Julian) 如果需填入資產，但資產為空，則聚焦到資產欄位
+  //     //     if (isAssetRequired && temporaryAssetList.length === 0) {
+  //     //       focusableElements[assetIndex]?.focus();
+  //     //     } else {
+  //     //       ToNext();
+  //     //     }
+  //     //     break;
+  //     //   case accountTitleIndex:
+  //     //     console.log('accountTitleIndex');
+  //     //     ToNext();
+  //     //     break;
+  //     //   default:
+  //     //     console.log('default');
+  //     //     focusableElements[0]?.focus();
+  //     //     break;
+  //     // }
+  //   },
+  //   [formRef, date, counterparty, isCounterpartyRequired, temporaryAssetListByUser]
+  // );
 
-  useHotkeys('tab', handleTabPress);
+  // useHotkeys('tab', handleTabPress);
 
   const dateRef = useRef<HTMLDivElement>(null);
   const counterpartyInputRef = useRef<HTMLInputElement>(null);
@@ -1079,21 +1079,23 @@ const NewVoucherForm: React.FC<NewVoucherFormProps> = ({ selectedData }) => {
             {t('journal:ADD_NEW_VOUCHER.VOUCHER_TYPE')}
             <span className="text-text-state-error">*</span>
           </p>
-          <button
-            id="voucher-type"
-            type="button"
-            onClick={typeToggleHandler}
-            className="relative flex items-center justify-between rounded-sm border border-input-stroke-input bg-input-surface-input-background px-12px py-10px hover:cursor-pointer hover:border-input-stroke-input-hover"
-          >
-            <p
-              className={`text-base ${isShowAnalysisPreview ? inputStyle.PREVIEW : 'text-input-text-input-filled'}`}
+          <div className="relative">
+            <button
+              id="voucher-type"
+              type="button"
+              onClick={typeToggleHandler}
+              className="flex w-full items-center justify-between rounded-sm border border-input-stroke-input bg-input-surface-input-background px-12px py-10px hover:cursor-pointer hover:border-input-stroke-input-hover"
             >
-              {isShowAnalysisPreview ? translateType(aiType) : translateType(type)}
-            </p>
-            <FaChevronDown size={20} />
+              <p
+                className={`text-base ${isShowAnalysisPreview ? inputStyle.PREVIEW : 'text-input-text-input-filled'}`}
+              >
+                {isShowAnalysisPreview ? translateType(aiType) : translateType(type)}
+              </p>
+              <FaChevronDown size={20} />
+            </button>
             {/* Info: (20240926 - Julian) Type dropdown */}
             {typeDropdownMenu}
-          </button>
+          </div>
         </div>
 
         {/* Info: (20240926 - Julian) Note */}
